More explicit time series construction
Separate constructors for everything, and code logic instead of a plain
lookup.
