--- static/app/views/dashboards/widgets/timeSeriesWidget/seriesConstructors/completeLineChartSeries.tsx
@@ -1,14 +1,0 @@
-import LineSeries from 'sentry/components/charts/series/lineSeries';
-
-import type {TimeseriesData} from '../../common/types';
-
-export function CompleteLineChartWidgetSeries(timeserie: TimeseriesData) {
-  return LineSeries({
-    name: timeserie.field,
-    color: timeserie.color,
-    animation: false,
-    data: timeserie.data.map(datum => {
-      return [datum.timestamp, datum.value];
-    }),
-  });
-}
