+++ static/app/views/dashboards/widgets/timeSeriesWidget/seriesConstructors/completeLineChartWidgetSeries.tsx
@@ -1,0 +1,14 @@
+import LineSeries from 'sentry/components/charts/series/lineSeries';
+
+import type {TimeseriesData} from '../../common/types';
+
+export function CompleteLineChartWidgetSeries(timeserie: TimeseriesData) {
+  return LineSeries({
+    name: timeserie.field,
+    color: timeserie.color,
+    animation: false,
+    data: timeserie.data.map(datum => {
+      return [datum.timestamp, datum.value];
+    }),
+  });
+}
